(refactor) Flytt mottak-client sammen med de andre, endre config
