+++ src/main/kotlin/no/nav/familie/ba/soknad/api/config/ApplicationConfig.kt
@@ -112,10 +112,9 @@ internal class ApplicationConfig {
     ): RestOperations {
         return RestTemplateBuilder()
             .interceptors(
-                bearerTokenClientHttpRequestInterceptor,
-                MdcValuesPropagatingClientInterceptor(),
                 consumerIdClientInterceptor,
                 apiKeyInjectingClientInterceptor,
+                MdcValuesPropagatingClientInterceptor(),
             )
             .build()
     }
