+++ src/main/kotlin/no/nav/familie/ba/soknad/api/controllers/SøknadController.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.ba.soknad.api.controllers
 
 import no.nav.familie.ba.soknad.api.domene.Kvittering
-import no.nav.familie.ba.soknad.api.integrasjoner.MottakClient
+import no.nav.familie.ba.soknad.api.clients.mottak.MottakClient
 import no.nav.familie.ba.soknad.api.util.TokenBehandler
 import no.nav.familie.kontrakter.ba.søknad.Søknad
 import no.nav.familie.kontrakter.felles.Ressurs
