+++ src/test/kotlin/no/nav/familie/ba/soknad/api/integrasjoner/ClientMocks.kt
@@ -4,6 +4,7 @@ import io.mockk.Runs
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
+import no.nav.familie.ba.soknad.api.clients.mottak.MottakClient
 import java.time.LocalDateTime
 import no.nav.familie.ba.soknad.api.clients.pdl.FAMILIERELASJONSROLLE
 import no.nav.familie.ba.soknad.api.clients.pdl.PdlBrukerClient
